- Launcher updated to 1.02: 	-> Possibility to use the default color. Thank-you to Rezca for report. 	-> GUI to exclude files from (in the option window). This allow to avoid your modded files to be overwritten during update. 	-> Drop support for Sinhala (unmaintened and less than 1% translated). This language was non-supported by the game. 	-> Add a function to reset the mod in the default state. - Installer (non included): 	-> Force installer to make a folder named "Rise of Mankind - A New Dawn" and to decompress into "Mods/" folder.

+++ checker/main.cpp
@@ -84,20 +84,27 @@ int main(int argc, char *argv[])
 
     // Check for existing installation
     QFile temp("updating");
-    if(temp.exists()){
+    QFile reset("reset");
+    if(temp.exists() || reset.exists()){
         f_check updater;
         updater.PrepareUpdate();
         QMessageBox msgBox;
         msgBox.setText(QObject::tr("There seems to be a problem with the previous update. Checking file again."));
         msgBox.exec();
-        QFile::remove("updating");
         #ifdef _WIN32
         system("taskkill /IM rsync.exe");
         #endif
         #ifdef __linux
         system("killall rsync");
         #endif
-        updater.ActionUpdate();
+        if(temp.exists()){
+            QFile::remove("updating");
+            updater.ActionUpdate();
+        }
+        if(reset.exists()){
+            QFile::remove("reset");
+            updater.ActionReset();
+        }
         return a.exec();
     }
